refactor: extract CardPressStyle into its own file
